refactor(bot): create separate packages for wynn views
