+++ fazcord/view/wynn_history/player_history_view.py
